Move custom admin setting to dedicated file
Otherwise, the `confs` models were missing from admin
